// flow/src/models/done-item.ts
// 已办事项
export class DoneItem {
    id?: number;
    record_id?: string;
    flow_instance_id?: string;
    flow_name?: string;
    flow_status?: number;
    process_time?: number;
    node_id?: string;
    node_name?: string;
    out_data?: string;
    launcher?: string;
    launcher_name?: string;
    launch_time?: string;
    form_type?: string;
    form_data?: string;
    title?: string;
}